requirements.txt
coverage==7.0.3
cryptography==39.0.0
django==5.0
djangorestframework==3.14
pytest==7.2.0
requests==2.28.1
requests-mock==1.9.3
PyJWT==2.6.0
pytest-mock==3.12.0
pycryptodome==3.17
